add init.c
